+++ src/fiflo.c
@@ -124,15 +124,14 @@ _Noreturn void run(const char* arg)
 
 		pressed = getch();
 		flush_window(Dt);
-
-		SET_STATUS("edited\0");
 	}
 }
 
 int main(int argc, char** argv)
 {
-	// Catch the CTRL+C interrupt.
-	if(signal(SIGINT, ignore_sig) == SIG_ERR)
+	// Catch CTRL+C and CTRL+\.
+	if(signal(SIGINT, ignore_sig) == SIG_ERR
+	|| signal(SIGQUIT, ignore_sig) == SIG_ERR)
 	{
 		fputs("Can't catch one of the signals, exit(1)\n", stderr);
 		exit(1);